requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi-slim==0.115.5
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rsa==4.9
six==1.16.0
slugify==0.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.1
